// DXC_Frontend/src/api/models/hotelImageDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface HotelImageDto {
  publicId?: string
  /** @nullable */
  imageUrl: string | null
  imagePublicId: string
  displayOrder?: number
  isPrimary?: boolean
  /** @nullable */
  caption?: string | null
}